+++ expressionlanguage/src/main/java/code/expressionlanguage/structs/IntStruct.java
@@ -17,12 +17,12 @@ public final class IntStruct extends NumberStruct {
 
     @Override
     public double doubleStruct() {
-        return value;
+        return (double)value;
     }
 
     @Override
     public float floatStruct() {
-        return value;
+        return (float)value;
     }
 
     @Override
